+++ packages/vercel-api-js/package.json
@@ -28,10 +28,15 @@
     "tsc": "tsc --noEmit",
     "build": "unbuild",
     "prepack": "unbuild",
-    "generate": "openapi-codegen vercel"
+    "generate": "rimraf ./src/generated && kubb generate"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/SferaDev/openapi-clients.git"
+  },
+  "devDependencies": {
+    "@kubb/cli": "^3.10.11",
+    "@kubb/core": "^3.10.11",
+    "openapi-utils": "workspace:*"
   }
 }
